Added ToDo application
simple to do application granting ability to type a task, delete input if needed, submit a task as well as mark task completed and delete tasks.

this is however, not scaleable which i will fix in the future, as well as add color classes as well within my data array as a object key for users to select and specify color codes for specific tasks.